Ensure that callback URLs are using HTTPs for remote hosts
